\name{qtlsim-internal}
\alias{which.correct.sub}
\alias{which.pos}

\title{Internal qtlsim functions}

\description{
  Internal qtlsim functions.
}

\usage{
which.correct.sub(infer, truth, within=1)
which.pos(vect, cs)
}

\details{
  These are generally not to be called by the user.
}

\author{Karl W Broman, \email{broman@wisc.edu}}

\keyword{internal}
